Custom data types for id/double pairs, refactor into model package
