+++ src/main/kotlin/ch/unibas/dmi/dbis/vrem/database/VREMReader.kt
@@ -5,7 +5,10 @@ import ch.unibas.dmi.dbis.vrem.model.exhibition.Exhibit
 import ch.unibas.dmi.dbis.vrem.model.exhibition.Exhibition
 import ch.unibas.dmi.dbis.vrem.model.exhibition.ExhibitionSummary
 import com.mongodb.client.MongoDatabase
-import org.litote.kmongo.*
+import org.litote.kmongo.eq
+import org.litote.kmongo.findOne
+import org.litote.kmongo.getCollection
+import org.litote.kmongo.projection
 
 /**
  * MongoDB reader for VREM.
