+++ src/main/kotlin/ch/unibas/dmi/dbis/vrem/generation/model/GenerationObject.kt
@@ -1,0 +1,8 @@
+package ch.unibas.dmi.dbis.vrem.generation.model
+
+enum class GenerationObject {
+
+    EXHIBITION,
+    ROOM
+
+}
