+++ src/main/kotlin/ch/unibas/dmi/dbis/vrem/generation/model/IdDoublePair.kt
@@ -1,0 +1,6 @@
+package ch.unibas.dmi.dbis.vrem.generation.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class IdDoublePair(val id: String, val value: Double)
