+++ src/main/kotlin/ch/unibas/dmi/dbis/vrem/generation/model/MetadataType.kt
@@ -1,0 +1,7 @@
+package ch.unibas.dmi.dbis.vrem.generation.model
+
+enum class MetadataType(val key: String) {
+
+    SOM_IDS("som.ids")
+
+}
